Application upgrade.

+++ application/managers/applications.py
@@ -57,24 +57,25 @@ class ApplicationManager:
                     dry_run=dry_run
                 )
         except HelmException:
-            logging.exception(
-                f'Application installation failed. Failed to install Helm chart "{chart.chart}". '
-                f'<Organization: id={user.organization.id}> <Template: id={template.id}>'
-            )
-            for chart in reversed(manifest_schema.charts[:index]):
-                try:
-                    await self.helm_manager.uninstall_release(
-                        organization=user.organization,
-                        context_name=context_name,
-                        namespace=namespace,
-                        release_name=chart.name
-                    )
-                except HelmException:
-                    logging.exception(
-                        f'Failed to remove release "{chart.name}" during clean up after application installation '
-                        f'failure. <Organization: id={user.organization.id}> <Template: id={template.id}>'
-                    )
-                    pass
+            if not dry_run:
+                logging.exception(
+                    f'Application installation failed. Failed to install Helm chart "{chart.chart}". '
+                    f'<Organization: id={user.organization.id}> <Template: id={template.id}>'
+                )
+                for chart in reversed(manifest_schema.charts[:index]):
+                    try:
+                        await self.helm_manager.uninstall_release(
+                            organization=user.organization,
+                            context_name=context_name,
+                            namespace=namespace,
+                            release_name=chart.name
+                        )
+                    except HelmException:
+                        logging.exception(
+                            f'Failed to remove release "{chart.name}" during clean up after application installation '
+                            f'failure. <Organization: id={user.organization.id}> <Template: id={template.id}>'
+                        )
+                        pass
             raise
         application_record = None
         if not dry_run:
@@ -98,7 +99,7 @@ class ApplicationManager:
         }
 
     async def update(
-        self, application: Application, organization: Organization, values: dict[str, dict], dry_run: bool = False
+        self, application: Application, values: dict[str, dict], dry_run: bool = False
     ) -> None:
         """
         Updates application release's values.
@@ -114,7 +115,7 @@ class ApplicationManager:
         results = {}
         for release_name, release_values in values.items():
             results[release_name] = await self.helm_manager.update_release(
-                organization=organization,
+                organization=application.organization,
                 context_name=application.context_name,
                 namespace=application.namespace,
                 release_name=release_name,
@@ -125,7 +126,99 @@ class ApplicationManager:
 
         return results
 
-    async def terminate(self, application: Application, organization: Organization) -> None:
+    async def upgrade(self, application: Application, template: TemplateRevision, dry_run: bool = False) -> None:
+        """
+        Upgrades application manifest to given template.
+        """
+        rendered_template = render_template(template.template, application.user_inputs)
+        new_template_schema = load_template(rendered_template)
+        old_template_schema = load_template(application.manifest)
+
+        charts_to_install = new_template_schema.chart_mapping.keys() - old_template_schema.chart_mapping.keys()
+        releases_to_remove = old_template_schema.chart_mapping.keys() - new_template_schema.chart_mapping.keys()
+        common_releases = old_template_schema.chart_mapping.keys() & new_template_schema.chart_mapping.keys()
+
+        install_results = {}
+        installed_charts = []
+        for release_name in charts_to_install:
+            chart = new_template_schema.chart_mapping[release_name]
+            try:
+                install_results[chart.name] = await self.helm_manager.install_chart(
+                    organization=application.organization,
+                    context_name=application.context_name,
+                    namespace=application.namespace,
+                    release_name=chart.name,
+                    chart_name=chart.chart,
+                    version=chart.version,
+                    values=chart.values,
+                    dry_run=dry_run
+                )
+                installed_charts.append(chart)
+            except HelmException:
+                if not dry_run:
+                    logging.exception(
+                        f'Application upgrade failed. Failed to install Helm chart "{chart.chart}". '
+                        f'<Organization: id={application.organization.id}> <Template: id={template.id}>'
+                    )
+                    for chart in reversed(installed_charts):
+                        try:
+                            await self.helm_manager.uninstall_release(
+                                organization=application.organization,
+                                context_name=application.context_name,
+                                namespace=application.namespace,
+                                release_name=chart.name
+                            )
+                        except HelmException:
+                            logging.exception(
+                                f'Failed to remove release "{chart.name}" during clean up after application upgrade '
+                                f'failure. <Organization: id={application.organization.id}> '
+                                f'<Template: id={template.id}>'
+                            )
+                            pass
+                raise
+
+        update_results = {}
+        for release_name in common_releases:
+            chart = new_template_schema.chart_mapping[release_name]
+            update_results[release_name] = await self.helm_manager.update_release(
+                organization=application.organization,
+                context_name=application.context_name,
+                namespace=application.namespace,
+                release_name=release_name,
+                chart_name=chart.chart,
+                values=chart.values,
+                dry_run=dry_run
+            )
+
+        for release_name in releases_to_remove:
+            chart = old_template_schema.chart_mapping[release_name]
+            try:
+                await self.helm_manager.uninstall_release(
+                    organization=application.organization,
+                    context_name=application.context_name,
+                    namespace=application.namespace,
+                    release_name=chart.name,
+                    dry_run=dry_run
+                )
+            except HelmException:
+                if not dry_run:
+                    logging.exception(
+                        f'Failed to remove release "{chart.name}" during application upgrade. '
+                        f'<Organization: id={application.organization.id}> <Template: id={application.template.id}>'
+                    )
+                pass
+
+        if not dry_run:
+            application.manifest = rendered_template
+            await self.db.save(application)
+
+        return {
+            'installed_releases': install_results,
+            'updated_releases': update_results,
+            'uninstalled_releases': list(releases_to_remove)
+        }
+
+    async def terminate(self, application: Application) -> None:
         """
         Terminates application. Removes all resources related with application.
         """
@@ -133,7 +226,7 @@ class ApplicationManager:
         for chart in manifest_schema.charts:
             try:
                 await self.helm_manager.uninstall_release(
-                    organization=organization,
+                    organization=application.organization,
                     context_name=application.context_name,
                     namespace=application.namespace,
                     release_name=chart.name
